src/main/java/nl/vollo/kern/VolloProperties.java
package nl.vollo.kern;

import org.springframework.boot.context.properties.ConfigurationProperties;
import org.springframework.boot.context.properties.EnableConfigurationProperties;
import org.springframework.stereotype.Component;

import java.time.Duration;

@Component
@EnableConfigurationProperties()
@ConfigurationProperties(prefix = "vollo.kern")
public class VolloProperties {

    static final String ISSUER = "Vollo";

    private final Jwt jwt = new Jwt();

    public Jwt getJwt() {
        return jwt;
    }

    public static class Jwt {
        private String issuer = ISSUER;
        private Duration expiration = Duration.ofDays(1);
        private Duration clockSkew = Duration.ofMinutes(5);
        private String secret = "secret";

        public String getIssuer() {
            return issuer;
        }

        public void setIssuer(String issuer) {
            this.issuer = issuer;
        }

        public Duration getExpiration() {
            return expiration;
        }

        public void setExpiration(Duration expiration) {
            this.expiration = expiration;
        }

        public Duration getClockSkew() {
            return clockSkew;
        }

        public void setClockSkew(Duration clockSkew) {
            this.clockSkew = clockSkew;
        }

        public String getSecret() {
            return secret;
        }

        public void setSecret(String secret) {
            this.secret = secret;
        }
    }
}
